requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
pycoingecko>=3.1.0
tabulate>=0.9.0
termcolor>=2.3.0
colorama>=0.4.6
python-dateutil>=2.8.2
pytz>=2023.3